+++ src/core/async/modules/proxy/promise.spec.js
@@ -9,7 +9,11 @@
 import Async from 'core/async';
 
 describe('core/async/modules/proxy `promise`', () => {
-	const $a = new Async();
+	let $a;
+
+	beforeEach(() => {
+		$a = new Async();
+	});
 
 	it('simple promise wrapper with the resolved promise', async () => {
 		expect(await $a.promise(Promise.resolve('All fine'))).toBe('All fine');
@@ -36,5 +40,151 @@ describe('core/async/modules/proxy `promise`', () => {
 		});
 
 		expect(await res).toBe('All fine');
+	});
+
+	describe('suspendPromise/unsuspendPromise`', () => {
+		it('suspending a promise by id', async () => {
+			let i = 0;
+
+			const promise = $a.promise(
+				new Promise((resolve) => setTimeout(resolve, 150))
+			);
+
+			promise.then(() => {
+				i++;
+			});
+
+			expect(i).toEqual(0);
+
+			$a.suspendPromise(promise);
+			await $a.sleep(200);
+			expect(i).toEqual(0);
+
+			$a.unsuspendPromise(promise);
+			expect(i).toEqual(1);
+		});
+
+		it('suspending a promise by a label', async () => {
+			let i = 0;
+
+			const
+				label = {label: 'foo'};
+
+			const promise = $a.promise(
+				new Promise((resolve) => setTimeout(resolve, 150)),
+				label
+			);
+
+			promise.then(() => {
+				i++;
+			});
+
+			expect(i).toEqual(0);
+
+			$a.suspendPromise(label);
+			await $a.sleep(200);
+			expect(i).toEqual(0);
+
+			$a.unsuspendPromise(label);
+			expect(i).toEqual(1);
+		});
+
+		it('suspending a promise by a group', async () => {
+			let i = 0;
+
+			const
+				group = {group: 'foo'};
+
+			const promise = $a.promise(
+				new Promise((resolve) => setTimeout(resolve, 150)),
+				group
+			);
+
+			promise.then(() => {
+				i++;
+			});
+
+			expect(i).toEqual(0);
+
+			$a.suspendPromise(group);
+			await $a.sleep(200);
+			expect(i).toEqual(0);
+
+			$a.unsuspendPromise(group);
+			expect(i).toEqual(1);
+		});
+	});
+
+	describe('mutePromise/unmutePromise`', () => {
+		it('suspending a promise by id', async () => {
+			let i = 0;
+
+			const promise = $a.promise(
+				new Promise((resolve) => setTimeout(resolve, 150))
+			);
+
+			promise.then(() => {
+				i++;
+			});
+
+			expect(i).toEqual(0);
+
+			$a.mutePromise(promise);
+			await $a.sleep(200);
+			expect(i).toEqual(0);
+
+			$a.unmutePromise(promise);
+			expect(i).toEqual(0);
+		});
+
+		it('suspending a promise by a label', async () => {
+			let i = 0;
+
+			const
+				label = {label: 'foo'};
+
+			const promise = $a.promise(
+				new Promise((resolve) => setTimeout(resolve, 150)),
+				label
+			);
+
+			promise.then(() => {
+				i++;
+			});
+
+			expect(i).toEqual(0);
+
+			$a.mutePromise(label);
+			await $a.sleep(200);
+			expect(i).toEqual(0);
+
+			$a.unmutePromise(label);
+			expect(i).toEqual(0);
+		});
+
+		it('suspending a promise by a group', async () => {
+			let i = 0;
+
+			const
+				group = {group: 'foo'};
+
+			const promise = $a.promise(
+				new Promise((resolve) => setTimeout(resolve, 150)),
+				group
+			);
+
+			promise.then(() => {
+				i++;
+			});
+
+			expect(i).toEqual(0);
+
+			$a.mutePromise(group);
+			await $a.sleep(200);
+			expect(i).toEqual(0);
+
+			$a.unmutePromise(group);
+			expect(i).toEqual(0);
+		});
 	});
 });