+++ src/core/async/CHANGELOG.md
@@ -9,6 +9,12 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v3.61.1 (2021-10-25)
+
+#### :bug: Bug Fix
+
+* Fixed bugs with clearing/muting/suspending promises by their IDs `modules/proxy`
+
 ## v3.50.0 (2021-06-07)
 
 #### :rocket: New Feature