+++ src/core/async/modules/proxy/CHANGELOG.md
@@ -9,6 +9,12 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v3.61.1 (2021-10-25)
+
+#### :bug: Bug Fix
+
+* Fixed bugs with clearing/muting/suspending promises by their IDs
+
 ## v3.30.3 (2021-01-28)
 
 #### :house: Internal